Migrate feeds_lang_disp().
